requirements.txt
aniso8601==3.0.2
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-filter==2.1.0
django-graphql-jwt==0.2.1
django-graphql-social-auth==0.1.4
django-heroku==0.3.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
idna==2.8
oauthlib==3.0.1
promise==2.2.1
psycopg2==2.8.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
urllib3==1.24.2
whitenoise==4.1.2
